Add UART transport (#9)
Modify transport file name

Rename project folder to _USB

Add app_colcon.meta to all projects

Add serial tests support

Modify transport file name on ThreadX

Modify transport file name on USB

Add serial project

Update submodules

Update submodules

Update projects on gitlab ci

Remove config file from src

Fix name

Revert config.h delete on UDP projects

Set flags to avoid warnings

Disable FreeRTOS lib warnings

Add missing functions declaration

Fix threadX time conversion and warnings

+++ e2studio_project_threadX/src/utils.h
@@ -4,7 +4,7 @@
 
 #include "hal_data.h"
 
-#define sleep_ms(X) tx_thread_sleep(((X / 1000.0) * TX_TIMER_TICKS_PER_SECOND))
+#define sleep_ms(X) tx_thread_sleep(((ULONG) ((X / 1000.0) * TX_TIMER_TICKS_PER_SECOND)))
 
 extern bsp_leds_t g_bsp_leds;
